requirements.txt
py-cord==2.4.1
firebase_admin
PyPDF2
python-dotenv